task_api/requirements.txt
flask>=3.0.0
gunicorn>=21.2.0
pydantic>=2.7.0
pyyaml>=6.0.0
